feat: fixing service
