+++ assets/svc.sh
@@ -13,7 +13,7 @@ UNIT_PATH=/etc/systemd/system/${SVC_NAME}
 TEMPLATE_PATH=$REDCARBON_AGENT_SERVICE_TEMPLATE
 IS_CUSTOM_TEMPLATE=0
 if [[ -z $TEMPLATE_PATH ]]; then
-    TEMPLATE_PATH=./bin/template/redcarbon.agent.service.template
+    TEMPLATE_PATH=./bin/templates/redcarbon.agent.service.template
 else
     IS_CUSTOM_TEMPLATE=1
 fi
